backend/requirements.txt
fastapi
uvicorn[standard]
httpx
selenium
openai
webdriver-manager
python-dotenv
pandas
google-cloud-firestore
